Some light code refactoring.
