+++ tfimm/models/__init__.py
@@ -1,3 +1,13 @@
 from .config import ModelConfig  # noqa: F401
-from .registry import register_model  # noqa: F401
+from .factory import create_model, create_preprocessing, transfer_weigths  # noqa: F401
+from .registry import (  # noqa: F401
+    is_model,
+    is_model_in_modules,
+    is_model_pretrained,
+    list_models,
+    list_modules,
+    model_class,
+    model_config,
+    register_model,
+)
 from .serialization import keras_serializable  # noqa: F401
